Add files via upload
"Given a function fn, return a memoized version of that function.
A memoized function is a function that will never be called twice with the same inputs. Instead it will return a cached value."

Kind of tricky but I figured it out. Just use a cache with JSON. Calculate the key with the arguments. If key exists, then return its value. Otherwise, calculate result of the function and store its value, using the key of the arguments.