Add files via upload
I'm embarrassed, but this problem has made me realize that for functions like `setTimeout`, where the first parameter is a function, that function can be an arrow function.
I know you can pass functions, but I struggle to see arrow functions as a function entity, so to speak.
Anyway, the only difference between the first and second functions is the first parameter. The first one receives a function that executes something (with a body), so it executes it (after waiting for the specified timeout). The second one receives a function, but it's empty, so it does nothing.

I've also learned that when you create a promise, JavaScript is actually providing you with a series of functions:
new Promise((resolve, reject) => {
    // resolve = function(value) { 
    //     // Correct
    // }
    // reject = function(error) { 
    //     // Error
    // }
})