Another promise-related one.
You need to create a function that returns a promise that resolves when all the promises it receives as parameters resolve. If one of the child promises is rejected, then the function returns the rejected parent promise, with the rejection reflected in the child promise. Basically, it's like using `Promise.All`.

Chaining `.then` and `.catch` is very useful.