First solution: mine Second solution: IA
I'm a little sleepy, don't judge me.

In the first solution, I go one by one, always adding the checked element to the subarray, and if I reach the specified size, I add that subarray to the result array and empty it. Before returning, I check if there are still elements in the subarray, and if there are, I add the subarray to the result (in which case the size of the last subarray is less than the chunk size).

The second solution uses `.slice` to iterate over the subarray by size increments. Much more elegant.

Great examples here: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Array/slice